+++ library/src/backend/device/device_math.cpp
@@ -992,13 +992,13 @@ void linalg::device_csrilu0_compute(csr_matrix& A, const csrilu0_descr* descr)
     }
 }
 
-void linalg::device_tridiagonal_solver(int                   m,
-                                       int                   n,
-                                       const vector<double>& lower_diag,
-                                       const vector<double>& main_diag,
-                                       const vector<double>& upper_diag,
-                                       const vector<double>& b,
-                                       vector<double>&       x)
+void linalg::device_tridiagonal_solver(int                  m,
+                                       int                  n,
+                                       const vector<float>& lower_diag,
+                                       const vector<float>& main_diag,
+                                       const vector<float>& upper_diag,
+                                       const vector<float>& b,
+                                       vector<float>&       x)
 {
     ROUTINE_TRACE("linalg::device_tridiagonal_solver");
