+++ library/src/backend/host/host_math.h
@@ -173,13 +173,13 @@ namespace linalg
 
     void host_csrilu0_compute(csr_matrix& A, const csrilu0_descr* descr);
 
-    void host_tridiagonal_solver(int                   m,
-                                 int                   n,
-                                 const vector<double>& lower_diag,
-                                 const vector<double>& main_diag,
-                                 const vector<double>& upper_diag,
-                                 const vector<double>& b,
-                                 vector<double>&       x);
+    void host_tridiagonal_solver(int                  m,
+                                 int                  n,
+                                 const vector<float>& lower_diag,
+                                 const vector<float>& main_diag,
+                                 const vector<float>& upper_diag,
+                                 const vector<float>& b,
+                                 vector<float>&       x);
 }
 
 #endif
